Migrate openinfoman interface to its own module
