--- config/default.json
@@ -1,15 +1,0 @@
-{
-  "openinfoman": {
-    "url": "localhost",
-    "port": 8984,
-    "path": "/CSD/documents.json",
-    "agent": false
-  },
-  "openhimApi": {
-    "username": "root@openhim.org",
-    "password": "password",
-    "host": "localhost",
-    "protocol": "https",
-    "port": 8080
-  }
-}
