+++ lib/index.js
@@ -5,35 +5,10 @@ var http = require("http");
 var fs = require("fs");
 const _ = require('lodash');
 
-var config = JSON.parse(fs.readFileSync("config/default.json"));
-var OpenHIM = require('./openhim.js');
+var OpenHIM = require('./openhim');
+const OpenInfoMan = require('./openinfoman')
 
-exports.getDocuments = getDocuments = function(callback) {
-  http.get(config.openinfoman, (res) => {
-    var data = "";
-    var resultArray = [];
-
-    res.setEncoding('utf8');
-
-    res.on('error', (err) => {
-      return callback(err);
-    });
-
-    res.on('data', (chunk) => {
-        data+=chunk;
-    });
-
-    res.on('end', () => {
-      var dataObj = JSON.parse(data);
-      for(var doc in dataObj) {
-        resultArray.push(doc);
-      }
-      callback(null, resultArray);
-    });
-  });
-}
-
-exports.createChannelConfigObject = createChannelConfigObject = function(name, callback) {
+exports.createChannelConfigObject = function (name, callback) {
   var channelConfig = _.cloneDeep(require('../resources/channelConfig.json'));
   channelConfig.name = name;
   channelConfig.urlPattern = "^/CSD/csr/"+ name +"/careServicesRequest.*$";
@@ -43,8 +18,6 @@ exports.createChannelConfigObject = createChannelConfigObject = function(name, c
   channelConfig.routes[0].port = 6000;
   callback(null, channelConfig);
 }
-
-module.exports = exports;
 
 exports.run = function(options, callback) {
   if (!callback) {
@@ -57,11 +30,13 @@ exports.run = function(options, callback) {
     url: options.openhim
   })
 
+  const openinfoman = OpenInfoMan(options.openinfoman)
+
   console.log('Fetching OpenInfoMan Documents...');
-  exports.getDocuments((err, docs) => {
+  openinfoman.getDocuments((err, docs) => {
     docs.forEach((doc) => {
       console.log(`Creating channel for ${doc}`);
-      createChannelConfigObject(doc, (err, channel) => {
+      exports.createChannelConfigObject(doc, (err, channel) => {
         openhim.registerChannel(channel, (err, res) => {})
       })
     })
